+++ zoo/src/main/scala/com/intel/analytics/zoo/friesian/python/PythonFriesian.scala
@@ -297,37 +297,12 @@ class PythonFriesian[T: ClassTag](implicit ev: TensorNumeric[T]) extends PythonZ
 
     var tmpDF = df
     for(col <- cols.asScala.toList) {
-      tmpDF = addValueSingleCol(tmpDF, col, mapBr, key, value)
+      tmpDF = Utils.addValueSingleCol(tmpDF, col, mapBr, key, value)
     }
 
     tmpDF
   }
 
-  def addValueSingleCol(df: DataFrame, colName: String, mapBr: Broadcast[Map[Int, Int]],
-                        key: String, value: String): DataFrame = {
-
-    val colTypes = df.schema.fields.filter(x => x.name.equalsIgnoreCase(colName))
-    val lookup = mapBr.value
-    if(colTypes.length > 0) {
-      val colType = colTypes(0)
-      val replaceUdf = colType.dataType match {
-        case IntegerType => udf((x: Int) => lookup.getOrElse(x, 0))
-        case ArrayType(IntegerType, _) =>
-          udf((arr: WrappedArray[Int]) => arr.map(x => lookup.getOrElse(x, 0)))
-        case ArrayType(ArrayType(IntegerType, _), _) =>
-          udf((mat: WrappedArray[WrappedArray[Int]]) =>
-            mat.map(arr => arr.map(x => lookup.getOrElse(x, 0))))
-        case _ => throw new IllegalArgumentException(
-          s"Unsupported data type ${colType.dataType.typeName} " +
-            s"of column ${colType.name} in addValueFeatures")
-      }
-
-      df.withColumn(colName.replace(key, value), replaceUdf(col(colName)))
-
-    } else {
-      df
-    }
-  }
 
   def mask(df: DataFrame, cols: JList[String], maxLength: Int): DataFrame = {
